requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
kiteconnect==4.2.0
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
streamlit==1.28.2
python-dotenv==1.0.0
pytz==2023.3
plotly>=5.18.0
google-generativeai>=0.3.2
